Async begone.
With Kludgine being just a renderer, the async nature of it was not as
appealing. With Gooey solving most of the problems I was trying to solve
with interior mutability (hierarchy and components), Kludgine no longer
needed a lot of its locking code. A lot of the async code in Kludgine
was to deal with locking structures within the layer that was meant to
also interact with your data. Gooey is going to aim to solve that union
in its own way, leaving Kludgine to focus on being a good renderer
primarily. This refactor is a step in the right direction, but there's
still mess left behind from some of the old design.

There's a couple more threads than there should be, but after a while, I
decided to focus on just finishing the task of ripping async out. And
now, it's late enough I just want to call this chunk of work done for
the day.