Reverting manual call to exit
The winit implementation on x11 still calls exit under the hood, but at least now I can't be blamed for it. Trying to diagnose crash-on-exit.

The last change is restoring the window closed notificaiton which trigger the app to exit, but doing it when the render thread exits rather than when we mark the window should close. This delay did not fix it.