Removed text wrapping
What Kludgine is becoming might be "Kludgine-core". As it stands, the
wrapping code is needed in Gooey in the cross-platform Rasterizer layer,
and there's no reason that code can't be re-integrated into Kludgine,
but it might require making an omnibus kludgine crate to do so.

For now, this branch is just going to continue focusing Kludine on its
rendering needs. The only reason for the stylecs dependency right now is
the shared Unit types for euclid.