Refactored scene size to u32
While this means size() calculates the scaled size each time, it allows
for lossless paths to be used when dealing with pixels.

Additionally, clipping now is ensured to be constrained to the scene's
size, whereas previously it went unchecked. This only becomes an issue
if you have inaccuracies in floating point numbers in most situations.