Updated more drawing calls to Drawable
Now all drawing calls should be flexible enough to allow any unit to be
passed in. This comes at a sacrifice of needing to specify the units in
some hard-coded situations. But, given that reducing the number of
floating point calculations done is a way to keep precision as high as
possible, the idea of being able to do pixel-alignment without
needlessly converting back to Scaled is worth it.