Fix CI freezing (#53)
In my haste to try to work on Gooey last night, I published a kludgine update without waiting for CI to finish. After both Kludgine and Gooey's CI timed out, I launched the investigation.

In the end there is a "fix" -- I think the old code (while amateurish with its usage of boxing futures and fusing, which are unnecessary here I now know) was technically more correct than what is happening here. I put in an issue for the wgpu developers to see: gfx-rs/wgpu#2266.

This also overhauls the CI workflows so that they're split. The dependencies have been switched to those that are in the wgpu repository's workflows.