Fixing breaking issue with rusttype 0.9.3's update
This was an unfortunate misconfiguration on my part due to how early
Kludgine was developed in my understanding of package conflicts. Because
I was depending on ttf-parser which rusttype also depends on, rusttype
updating from ttf-parser 0.14 to 0.15 between 0.9.2 and 0.9.3 is
considered a breaking change *only for some projects*.

The safest thing to do when importing dependencies that another crate
also depends on is to pin the version of the crate that you're sharing
dependendicies with.

Because I don't expect there are many users of Kludgine, and I doubt
those users are independently using `rusttype` rather than just using
the built-in font support, I'm treating this as a non-breaking change.

However, I really should pick a policy and document it in the Readme.