Texture coordinates are now integers
There is a lot of random cleanup in this commit, but the only real
change in functionality is switching from inputting precomputed f32
fractional coordinates to unsigned pixels for the texture coordinates.

This has a couple of major benefits: resizing textures no longer breaks
texture coordinates, and fractionally extracting from a texture is no
longer possible.

Subpixel rendering may be desired, however. I'm currently debating
whether to always pixel align or offer some sort of integer-based
subpixel subdivisions. With the modern movement being to move to higher
DPI screens, in general, the importance of subpixel rendering dimenishes
and the sharpness of pixel aligned graphics shines.

Ultimately, I'm sure this will be tweaked once text rendering is
supported, as that's the situation where subpixel rendering seems like
it should be important.