Cleanup + Clippy + Rustfmt
This commit should have no functionality changes except for edge case
handling. The public interface has been pruned so that the types listed
by rustdoc are the minimal types, and a lot of this was done by
consolidating the wgpu pipeline logic into a new pipeline module.

Next, I added safety comments for unsafe usage. There's only a few
usages for wgpu and the rest are empty bytemuck::Pod implementations.

The rest of the changes are the fallout from enabling clippy::pedantic
and adding a rustfmt.toml.