Fixing scaling factor
When run on my Mac which has a scaling factor of 2.0, the
supposed-to-be-1-inch square was super tiny instead.

Turns out I had messed up about how the numbers were being combined when
bytemucking the Ratio. Stamping two u16s together seems harmless, but it
actually inverts the byte order on little endian architectures. Thus,
the numerator was becoming the denominator.

By manually creating the u32 with bitshifts, the issue goes away and the
correct scaling factor is applied.

Finally, the shapes example has been updated. The blue triangle's
"outer" edge is now 96 pixels. This was chosen because at 96 dpi, the
triangle should be 1 inch in size. This allows for the scaling factor to
be observed more easily and see which direction the scaling is applied
in.