Sharing cached glyphs between multiple renders
This commit makes glyph caching be shared between multiple calls to
prepare_text. Kludgine::next_frame needs to be called between frames to
clean up unused glyphs. This manual clean up process allows a fuzziness
in the reference counting so that if a glyph is freed and then redrawn
within the same frame, it will not be removed and re-added to the cache.

While this technically can lead to more texture utilization, the
alternative is that poorly written code could have a harsh penalty in
performance.

Lastly, I wanted to see if I had to do anything special to support RTL
text (#55). I added a second line to the text example which included RTL
text, and it worked perfectly without any additional changes needed.