Text is now more efficient and flexible
Previously text was rendered internally using a Rendering structure,
because it supported rendering multiple commands. However, a Rendering
is very inflexible, because it supports rotating, scaling, and
translation on all of its internal drawing commands.

This commit makes PreparedText use PreparedGraphic internally, and now
PreparedGraphic supports multiple commands, all which are executed using
the parameters of the draw call at the time of rendering.

The other benefit of this refactor is I've batched the drawing calls.
The text example is drawn in 3 gpu calls now: the first set of text, the
emoji, and the second set of text.