Drawing to textures is supported now
The flow isn't very nice, but it works. I struggled a lot with
the 'pass lifetime management. This works, but I haven't been
able to encapsulate as much of the wgpu behavior as I would like
-- almost to the point that I'm not sure the encapsulation I've
done is all that helpful.

It's easier to work from a place of functioning code -- so I'm
committing it as is and hoping to refine it in the days to come.