Refactored rendering encapsulation
The `Kludgine` type now returns a Frame from next_frame, which is how a
RenderingGraphics can be constructed now. RenderingGraphics no longer
accepts a pass, but rather a pass descriptor, and Frame contains the
command encoder. This particular structuring of flow fits the lifetime
requirements of the RenderPass, and it makes the offscreen example a
little cleaner.