Added Pong example
It has some of the classical edge cases due to how simple it is, such as
a ball hitting multiple times inside of the paddle. But the goal wasn't
to write a *good* game, just to have an example showing how it all comes
together.

In the process I realized that the Renderer::draw_text function didn't
take a color. I think I was planning on having a setting for text color,
but I prefer being able to pass it in during the draw call.