Added clipping to Graphics/Drawing
`Drawing` needed to support clipping for my vision of how I want to render
in Gooey. As I dug in, I realized that while Graphics doesn't need a
clip rect for most purposes, the requirement that `Drawing` have a clip
rect introduced a conundrum for the Graphics/RenderingGraphics phases --
how can nested structures use both a prepare()/render() phase with
clipping support without forcing library authors to pass along the clip
rect to the prepare phase?

This is further complicated by the clip rects being relative as you nest
them. So, the best option was to generalize clipping and implement it
across all the context types.