Refactored clipping
In Gooey, I needed a way to clip without the lifetime issues, so I've
introduced a push/pop clip API. In the process, I realized I could unify
a lot of the clip handling.

In this process I also discovered that setting a scissor rect to an
empty rect causes wgpu to have a validation error. So, now any time the
clipping rect is empty, the drawing calls are skipped before even
sending them to the GPU.