Fixed index buffer update issues
I knew that theoretically the index buffer could be updated with a
stride that was not even, but it took a while to actually hit that edge
case. DiffableBuffer now tries to fix alignment by adding one extra
element to copy, otherwise it reallocates.

Vertex is 4-byte aligned, and index buffers are u16s, so this strategy
should mean that delta updates are always possible.