Measuring wrapped text, new test/example
The main changes here are adding a new width parameter to many text
drawing/measuring calls. This causes the text to be wrapped at the
provided width. These functions take so many parameters, they should be
refactored into an options type.

The other example, outline, shows drawing a 1px stroked rect in both the
window's contents and a clipped context. Before changing floating points
in figures to round on conversion, this example was not drawing the
top or left sides of the stroked rectangle. This was due to
lyon_tesselation generating coordinates that were ever-so-slightly under
1.0, and thus being truncated to 0 leading to 0-width rectangles for
  those lines.