Blind fix attempt: u32 in push constant?
In my earlier searching for a likely cause to #93, I found that it might
be stemming from push constants being converted automatically to
uniforms by wgpu when the backend doesn't support push constants.
Looking at the types on the push constants structure, the only one that
didn't seem to be supported is u32.

Unfortunately, I haven't created a setup where I can reproduce this
issue yet, so this is just a blind attempt while I get one set up.