Fix: Text rendering subpixel alignment
Due to Gooey using ::Center everywhere, I couldn't ever figure out the
right adjustment to perform for the subpixel offset. I did not realize
that cosmic text was handling subpixels on the texture so that the extra
calculation doesn't need to be done when blitting the glyph.

Fixing TextOrigin::Center to round its calculation to the nearest whole
pixel was the real solution to making subpixel alignment be correct.