Attempt workaround for broken scissoring
Refs khonsulabs/gooey#93

This appears to work around the issue on the GL backend. I cannot
currently reproduce this behavior in a simpler example. This should have
no effect, as the documentation for set_scissor_rect states that it
should only affect future drawing calls... but clearly there are no
future drawing calls being done at the point of dropping the graphics.