Fixed Srgb handling on shape drawing
When starting to try to test Cushy's UI with screen captures, I realized
the color output was significantly off. I remembered that I had picked a
very simple srgb to linear conversion algorithm I had found that I could
implement in wgsl easily.

This commit removes that simplistic conversion with a proper one powered
by palette. This has to be done before sending the color data to the
shader.

With this, I'm able to start writing unit tests that test pixel colors
in Cushy.