Updated wgpu to 22.0.0
The lifetime removal didn't really impact the design of Kludgine due to
the command encoder lifetime still being bound on the structure stored
in RenderingGraphics.

We can opt out of this lifetime dependency, but I'm not really sure the
implications of doing it, so the path of least resistance at the moment
is just this minimal change.