Fixing `Drawing` with clip rects
The translation for clipping being applied in Drawing::render was
incorrect after the changes in the last commit. There are two separate
clip stacks in the case of rendering a `Drawing`: the clip stack that
happened while rendering the `Drawing`'s frame and the clip stack on the
graphics that is currently rendering the prepared `Drawing`. We need to
ensure that the origins of both clip rects are being applied to all
opreations, and this set of changes does that.

This change also fixes the logic around when to skip sending drawing
calls when the calls are fully clipped. Previously only the first
clipped drawing call would be ignored, and because of panics inside of
set_scissor_rect, the clip rect was never applied.