Updating CI + adding RenderOperation::finish
When thinking of how someone might add a custom batched renderer to
replace the shape renderer entirely, I realized there needed to be one
last chance for the operations to do something before the render pass
begins. The Drop implementation of Renderer is where the vertex buffers
for the shape operations are all uploaded to the GPU -- and now
RenderOperation::finish is also invoked when finalizing the rendering
allowing for similar operations to be batched.